Sicredi/app/src/main/java/com/example/sicredi/viewmodel/EventViewModel.kt
package com.example.sicredi.viewmodel

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import com.example.sicredi.data.OperationCallback
import com.example.sicredi.model.Event
import com.example.sicredi.model.EventRepository

class EventViewModel(private val repository: EventRepository) : ViewModel() {

    private val _events = MutableLiveData<List<Event>>().apply { value = emptyList() }
    val events: LiveData<List<Event>> = _events

    private val _isViewLoading = MutableLiveData<Boolean>()
    val isViewLoading: LiveData<Boolean> = _isViewLoading

    private val _onMessageError = MutableLiveData<Any>()
    val onMessageError: LiveData<Any> = _onMessageError

    private val _isEmptyList = MutableLiveData<Boolean>()
    val isEmptyList: LiveData<Boolean> = _isEmptyList

    init {
        loadEvents()
    }

    fun loadEvents() {
        _isViewLoading.value = true
        repository.fetchEvents(object : OperationCallback<Event> {
            override fun onError(error: String?) {
                _isViewLoading.value = false
                _onMessageError.value = error
            }

            override fun onSuccess(data: List<Event>) {
                _isViewLoading.value = false
                if (data.isNullOrEmpty()) {
                    _isEmptyList.value = true

                } else {
                    _events.value = data
                }
            }
        })
    }

}